remove un-needed casts

+++ java/src/jmri/jmrix/ieee802154/xbee/XBeeReply.java
@@ -35,7 +35,7 @@ public class XBeeReply extends jmri.jmrix.ieee802154.IEEE802154Reply {
         xbresponse=new com.rapplogic.xbee.api.GenericResponse();
         byte ba[]=jmri.util.StringUtil.bytesFromHexString(s);
         int ia[]=new int[ba.length];
-        for(int i=0;i<ba.length;i++) ia[i]=(int)ba[i];
+        for(int i=0;i<ba.length;i++) ia[i]=ba[i];
         xbresponse.setRawPacketBytes(ia);
     }
